@@ -2,58 +2,61 @@
   "name": "next-clean-architecture",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "build": "next build",
-    "dev": "run-p dev:*",
+    "dev": "run-p 'dev:*'",
     "dev:next": "next dev",
     "dev:hyper": "npx hyper-nano --domain=next-clean-architecture --experimental --data --cache",
-    "format": "prettier --write .",
+    "format": "prettier '.' --write",
     "lint": "next lint",
-    "prepare": "husky install",
+    "prepare": "husky",
     "staged": "lint-staged --allow-empty",
     "start": "next start",
     "test": "vitest",
     "test:ci": "vitest --coverage"
   },
   "dependencies": {
-    "@graphql-tools/schema": "^9.0.12",
-    "cuid": "^2.1.8",
-    "dataloader": "^2.1.0",
-    "graphql": "^16.6.0",
+    "@graphql-tools/schema": "^10.0.4",
+    "cuid": "^3.0.0",
+    "dataloader": "^2.2.2",
+    "graphql": "^16.9.0",
     "graphql-helix": "^1.13.0",
     "graphql-tag": "^2.12.6",
-    "hyper-connect": "^0.6.0",
-    "next": "13.1.1",
-    "ramda": "^0.28.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "undici": "^5.14.0",
-    "zod": "^3.20.2"
+    "hyper-connect": "^0.12.0",
+    "next": "14.2.5",
+    "ramda": "^0.30.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "undici": "^6.19.5",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@anatine/zod-mock": "^3.8.2",
-    "@faker-js/faker": "^7.6.0",
-    "@types/node": "18.11.18",
-    "@types/ramda": "^0.28.20",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
-    "@vitest/coverage-c8": "^0.27.0",
-    "autoprefixer": "^10.4.13",
-    "dotenv": "^16.0.3",
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
-    "eslint-config-prettier": "^8.6.0",
-    "husky": "^8.0.3",
-    "hyper-nano": "^1.3.1",
-    "lint-staged": "^13.1.0",
+    "@anatine/zod-mock": "^3.13.4",
+    "@faker-js/faker": "^8.4.1",
+    "@types/node": "22.0.2",
+    "@types/ramda": "^0.30.1",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "@vitest/coverage-c8": "^0.33.0",
+    "@vitest/coverage-v8": "^2.0.5",
+    "autoprefixer": "^10.4.19",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.5",
+    "eslint-config-prettier": "^9.1.0",
+    "husky": "^9.1.4",
+    "hyper-nano": "^3.0.0",
+    "lint-staged": "^15.2.7",
     "markdown-toc-gen": "^1.0.1",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.2",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "tailwindcss": "^3.2.4",
-    "typescript": "4.9.4",
-    "vite-tsconfig-paths": "4.0.3",
-    "vitest": "^0.27.0"
-  }
+    "postcss": "^8.4.40",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "tailwindcss": "^3.4.7",
+    "typescript": "5.5.4",
+    "vite-tsconfig-paths": "4.3.2",
+    "vitest": "^2.0.5"
+  },
+  "packageManager": "yarn@4.3.1"
 }
