@@ -1,4 +1,5 @@
-module.exports = {
+// eslint-disable-next-line import/no-anonymous-default-export
+export default {
   content: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
   plugins: []
 }
