@@ -1,7 +1,8 @@
+// eslint-disable-next-line import/no-anonymous-default-export
 module.exports = {
   printWidth: 100,
   singleQuote: true,
   semi: false,
   trailingComma: 'none',
-  plugins: [require('prettier-plugin-tailwindcss')]
+  plugins: [require.resolve('prettier-plugin-tailwindcss')]
 }
